requirements.txt
asgiref==3.9.2
black==25.9.0
click==8.3.0
colorama==0.4.6
Django==5.2.7
django-crispy-forms==2.4
djangorestframework==3.16.1
h11==0.16.0
isort==7.0.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
python-decouple==3.8
pytokens==0.1.10
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.37.0
